@@ -3,33 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.17",
     "@tanstack/react-query": "^4.32.6",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.0",
     "@types/node": "^16.11.19",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^17.0.65",
     "@types/styled-components": "^5.1.19",
     "axios": "^0.24.0",
     "cors": "^2.8.5",
     "dotenv": "^12.0.3",
     "express": "^4.17.2",
-    "framer-motion": "^5.5.6",
+    "framer-motion": "^5.6.0",
     "gh-pages": "^3.2.3",
     "nodemon": "^2.0.15",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^7.24.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
-    "react-query": "^3.34.7",
-    "react-router-dom": "^6.7.0",
+    "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.0",
-    "styled-components": "^5.3.3",
+    "styled-components": "^5.3.11",
     "typescript": "^4.5.4",
     "web-vitals": "^2.1.3"
   },
@@ -58,6 +52,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^18.2.7",
     "@types/react-helmet": "^6.1.5"
   }
 }
