@@ -1,5 +1,4 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
 import NotFound from './page/NotFound';
 import Home from './page/Home';
@@ -7,6 +6,7 @@ import Tv from './page/Tv';
 import Movie from './page/Movie';
 import Search from './page/Search';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 
 const router = createBrowserRouter([
@@ -23,11 +23,13 @@ const router = createBrowserRouter([
   },
 ]);
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
     <RouterProvider router={router} />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 export default App;
