@@ -1,9 +1,9 @@
-import { QueryClient, QueryClientProvider } from 'react-query';
 import { Outlet, useLocation } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
 import Header from './Components/Header';
 import { ApiProvider } from './context/ApiContext';
 import { theme } from './theme';
+import { QueryClientProvider, QueryClient } from '@tanstack/react-query';
 
 const queryClient = new QueryClient();
 
