@@ -1,4 +1,4 @@
-import { useQuery } from 'react-query';
+import { useQuery } from '@tanstack/react-query';
 import { useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import { GetResult } from '../api/api';
@@ -35,7 +35,7 @@ function Search() {
   );
 
   const loading = searchMovieLoading || searchTVLoading;
-  console.log(location.pathname);
+
   return (
     <Wrapper>
       {loading ? (
