@@ -1,5 +1,5 @@
 import { useEffect } from 'react';
-import { useQuery } from 'react-query';
+import { useQuery } from '@tanstack/react-query';
 import styled from 'styled-components';
 import { GetResult } from '../api/api';
 import VideoDetail from '../Components/VideoDetail';
